Revert "[docker-teamd]: Manage teamd and teamsyncd processes with supervisor (#1137)"
This reverts commit a6edef2fa52f5b7771cc5751cde1a190d96ffa50.

The reason to revert this commit is that it breaks the current nightly test as
no port channel interfaces are get created after boot. teamd failed to start and
complained about 'Cannot allocate memory' possibly due to nlmsg_alloc function
failure.

Will revert this commit to investigate it further before moving to supervisor.

Signed-off-by: Shu0T1an ChenG <shuche@microsoft.com>