[device/dell] Update Shared headroom values
This commit updates the shared headroom value for z9100
T0 and T1 profile based on the ratio. The ratio is derived
from the 40G broadcom recommended XL Sheet.
Unit tested the above configuration by dumping the registers
for the shared headroom. The shared headroom values in registers
reflects the correct new values for both T1 and T0.

Signed-off-by: Harish Venkatraman <Harish_Venkatraman@dell.com>