[Dell] z9100 port qos & buffer changes of 0330 to master
DO NOT MERGE. Please merge after or along with PR 2239
https://github.com/Azure/sonic-buildimage/pull/2239

This commit has the forward porting of changes from 20180330
to master. Unit tested by checking the broadcom registers
for the values. The values in hardware reflect the correct values
except for TC to PFC priority group map values. Master branch values
for TC to PFC priority group map for both z9100 and s6100 in incorrect,
I have a build of Oct 12 master and the values are correct in
that build.

Signed-off-by: Harish Venkatraman <Harish_Venkatraman@Dell.com>