Filesystems - eMMC wearout mitigation
Some platform's flash is eMMC, this can be ok for some NOS but SONiC is
very write intensive, especially logging. This patch aims to
prolong the life of these eMMC flash drives by moving the /var/log
filesystem from a loop device to a tmpfs RAMdrive

The patch checks the kernel boot variables for the presence of an MMC
module string and modifies the initramfs build to mount the /var/log
filesystem as a ramdisk. The check is crude and has only been tested on
our only known affected platform, we are not sure it may work on all
eMMC based flash storage.

Caveat: you lose all stored logs at each reboot on eMMC platforms so
configure your rsyslogs to log outside the box

Signed-off-by: Michel Moriniaux <m.moriniaux@criteo.com>