[platform]: Added exceptions handling for BFN syseeprom and psuutil (#3342)
Added exceptions handling for BFN psuutil.py and eeprom.py which are raised when syseepromd and psud try to connect to the BFN thrift server which is not up yet. Now the exceptions backtrace is not logged to the syslog. Also psud doesn't exit on system bootup due to uncaught exception.

Signed-off-by: Vitaliy Senchyshyn <vsenchyshyn@barefootnetworks.com>