System recovery when syncd crashes When syncd gets terminated unexpectedly system  goes into unstable state. And system stays in that state unless a reboot is triggered by the user. It doesn't get receovered  by iteslef.  This is the problem getting addressed  with this change.
When syncd process crashes, we cannot restart syncd alone as it has dependency on orch-agent.
So same mechanism when SAI API call failure happens will be used here. syncd process state is monitored and when it crashes,
shutdown notification is been sent to orch-agent which eventually result in SWSS restart and syncd restart, to recover the system.