Workaround for the boot problem from Golden bios
There are primary and secondary bios in qfx5210 platform.
There is a problem which prevents the OS booting from the
secondary bios when the OS was installed using primary bios.
Secondary bios fails to detect the UEFI partition. Right now
the workaround is to have a folder structure /EFI/BOOT/BOOT64x.efi

Signed-off-by: Ciju Rajan K <crajank@juniper.net>