[init_cfg.json.j2] remove hardcode in init_cfg.json.j2
In order to avoid some missed container/feature listed in
init_cfg.json.j2 in FEATURE and CONTAINER_FEATURE tables I changed the
way optional feature are defined. With this change it is declared by
feature makefile whether this feature is optional or not. As for
CONTAINER_FEATURE table, the configuration is generated based on all
installable dockers.

Signed-off-by: Stepan Blyschak <stepanb@mellanox.com>