To change error log condition in is_port_up()
For lldpmgrd, listen for changes to the PORT table in the CONFIG_DB and APP_DB
To handle alias/description config change.
To check if port is up or down by looking into the oper-status for the port in PORT TABLE in the App DB.
Don't find it in the App DB, will log error.
During initializing, it is possiable that port change in CONFIG_DB and not ready in APP_DB.