[sonic-host-service]: Add SONiC Host Services infrastructure
When SONiC is configured with the management framework and/or telemetry
services, the applications running inside those containers need to
access some functionality on the host system. The following is a
non-exhaustive list of such functionality:

* Image management
* Configuration save and load
* ZTP enable/disable and status
* Show tech support

The host service is a process that listens for requests via D-Bus. It
will then service those requests and send a response back to the
requestor.

Signed-off-by: Nirenjan Krishnan <Nirenjan.Krishnan@dell.com>