Fix config qos reload on Multi-asic platfroms. The sort Function on Portname is not able to understand Backplance Interface name.
Fix is to to remove sorting from j2 template as this is not needed
as all Ports are dervice port_config.ini where ports are listed
in fixed static sort order.

Also we do not sort port in buffer_config.j2 so made logic similar to
that and keep it simple instead of enhancing sort function.

Signed-off-by: Abhishek Dosi <abdosi@microsoft.com>