[make init] update submodule's remotes
This change is to address an issue where when using a long standing
repo to build image, the submodule could become out-dated with their
remotes so that make init is no longer to bring all submoudles to
the latest. The consequence is weird build failures.

Updatign all submodule's remotes addres the issue with the cost of
make init runs a bit longer.

Signed-off-by: Ying Xie <ying.xie@microsoft.com>