[hostcfgd] Configure service auto-restart in hostcfgd.
Before this change, a process runnning inside every SONiC
container dealt with FEATURE table 'auto_restart' field and
depending on the value decided wether a container has to be
killed or not. If killed service auto restart mechanism
restarts the container. This change moves the logic from
container to the host daemon - hostcfgd.

* hostcfgd refactoring - move feature handling in another class.
* override systemd service Restart= setting from hostcfgd.
* remove code that deals with FEATURE table from supervisor-proc-exit-listener.
* remove default systemd Restart=always.

Signed-off-by: Stepan Blyshchak <stepanb@nvidia.com>