[timers] Fix timer units.
'Conficts=' in timer units caused a problem when
starting a timer causes restart of a service.
Instead, 'PartOf=' is a better choice.
'PartOf=' will make the timer stop and restart togather
with the service the timer is bound to, otherwise the timer
OnUnitActiveSec expires imidiatelly when the service stops
and starts it.

Signed-off-by: Stepan Blyshchak <stepanb@nvidia.com>