[Arista 7050cx3] add dummy MMU configurations for Arista-7050CX3-32S-C32
These MMU configuraions are copied from another platform and
guaranteed to be incorrect for hwsku Arista-7050CX3-32S-C32.

Adding them so that we have A MMU configuration and the system
won't through a whole bunch of errors and leave MMU unconfigured.
The correct MMU configuration will come later.

This configuration is definitely not suitable for testing system
performance or QoS behavior.

Signed-off-by: Ying Xie <ying.xie@microsoft.com>