Merged PR 3745506: [Monit] Use the new syntax of Monit configuration file in ACMS container
Initially the syntax of checking the exit status of script monitored by Monit  is:

`if status != 0 for 5 times within 5 cycles then alert`.

Such syntax will fire an error message in syslog immediately if a process was not running, but the expectation is the error message should be fired if the process was not running in continuous 5 times within 5 cycles/minutes. So the new syntax introduced the keyword `alert repeat` which can achieve this purpose.

Signed-off-by: Yong Zhao <yozhao@microsoft.com>