eBPF library support for SONiC applications
It provides a generic library support for filtering the netlink message
using eBPF based packet filtering.
Linux netlink messaging system is a broadcast domain, every netlink
message is broadcasted to all the subscribed appplications regardless
whether the application is interested in it or not.
In a scaled scenario, more than 1 million message is being broadcasted
across applications.
This framework helps to filter the unwanted netlink message at the
kernel space itself and eliminates the unnecessary passing of message to the user space
application and then it get dropped. This improves the system
performance and scale the system to higher level.