Better handle tmpfs in boot0
Avoid doing memory to memory copy for images that are already in a tmpfs.

The current implementation does not detect properly if the swi is
already in a tmpfs. It now checks for tmpfs or ramfs before trying to
mount a new tmpfs and moving the swi.

The cleanup method has been updated accordingly and will only cleanup
the mount path if it's below /tmp/ as to not affect user mounted paths.