[sysctl] Increase hung_task_timeout_secs to 300
Depending on the performance characteristics of a given hardware platform,
it's possible to exceed the default 120 second kernel timeout during I/O
intensive operations like image installation. This risk increases as image
size continues to increase. So, we need to increase the timeout so that we
don't encounter kernel panics on devices with lower disk throughput.

Signed-off-by: Danny Allen <daall@microsoft.com>