[linkmgrd] Monitor MUX State When Link State Is Down
It is possible that link goes down on both ToR, then up on an active ToR
and then down again. This combination would result in tunnel being setup as
stanby on both ToR which would create a loop. This PR adds timer to the link
down case and update the tunnel when the MUC changes.

singed-off-by: Tamer Ahmed <tamer.ahmed@microsoft.com>