Update hostcfgd to disable unused services for fabric asics
Some services are unnecessary for fabric asics like bgp, lldp, teamd.
We so disable them. But it could be that there are some mixed asics
NPU and fabric in the same system, so we should just disable only
for fabric asics.

This commit is to disable those services only fabric asics.

Signed-off-by: ngocdo <ngocdo@arista.com>