Merged PR 4251943: [rsyslog.conf.j2] No longer hardcode production syslog server IP; Add square brackets around IP addr
Production syslog server (10.20.6.16) should be present in all minigraphs, so there should no longer be a need to hardcode it.

Hardcoding it here also can cause issues, as lab devices have the potential to spam the production syslog server and overload it.

Also add square brackets around IP address to support IPv6 addresses. This change was added in public a while ago, but it appears the internal patch was never updated to reflect it.