[arista] Add DPB for Arista 7050 QX32
This change introduces dynamic port breakout (DPB) for Arista 7050 QX32
model as a new SKU.

The breakout configuration allowed is the same as in mainline Arista
EOS, i.e. 24 first ports are allowed to be used in 4x10G in addition to
the default 40G mode. The last 8 ports are fixed to 40G. This is due to
ASIC limitations of a total of 104 max ports.

Signed-off-by: Christian Svensson <blue@cmd.nu>