[teamd/patch] do not send LACP PDU if link is down
Otherwise the lacp_port_set_state() returns error when called from
lacp_port_link_update() and lead to not up-to-date cache in
lacp_port->__link_last. This can happen when member flapping happens.
When this issue happens the condition comparing lacp_port->__link_last
to current linkup in lacp_port_link_update() will not pass and member
will never get out of deselected state unless it is flapped again.

This is done as a fix in patch maintained by SONiC but planned to be
upstreamed in libteam as well. Currently we cannot simply propose this
fix in libteam and use libteam master as of important bug fix revert in
jpirko/libteam 61efd6de2fbb8ee077863ee5a355ac3dfd9365b9.

Signed-off-by: Stepan Blyschak <stepanb@nvidia.com>