Merged PR 4715846: Invalidate RX Sequence No When Sending Heartbeats
Update RX Sequence No When Sending Heartbeats

For a server that is not responding to HBs, it is poosible to report
Active link incorrectly when 16-bit sequence number wraps around.
The issue was hidden when state change required 3 consecutive events
of the same type or lack of. With 1 event for Active/Standby transition,
the issue is visible.

singed-off-by: Tamer Ahmed <tamer.ahmed@microsoft.com>