[multiasic][voq][bgpconf] Fix for different router id for iBGP
For multiasic, the back end asics use ip addresss of Loopback4096 for
BGP router id. In VOQ multi-asic chassis there are no back end asics.
All the asics are front end and the iBGP connections are established via
Ethernet-IB of asics. Since these asics are not designated as BackEnd,
the ip address of interface Loopback0 is used as BGP router id. Since
the ip address of Loopback0 is same for all the asics, same router id is
used for voq iBGP configurations and hence the iBGP connections are not
established. To fix this, we need to use the ip address of Loopback4096 for
the front end asics also in multi-asic VOQ systems.

Changes are done to implement this fix.

Signed-off-by: vedganes <vedavinayagam.ganesan@nokia.com>