Pass ASIC vendor information to swss docker as docker level environment variable
Motivation:
Recently, the reserved buffer of admin-down ports is going to be reclaimed.
However, the way to do this differs among vendors.
We need to find a way to pass vendor information to swss docker.
Implementation:
Fetch the ASIC vendor information when the docker is created and pass it to docker as ASIC_VENDOR

Signed-off-by: Stephen Sun <stephens@nvidia.com>